rename purge to deleteAll
