+++ src/schema/deleteAll.ts
