+++ src/schema/journey.test.ts
@@ -396,9 +396,9 @@ describe('schema', () => {
     return deleteClass(client, newClass.class);
   });
 
-  it('purges the schema of all data', () => {
-    const newClass: any = newClassObject('LetsPurgeThisClass');
-    const newClass2: any = newClassObject('LetsPurgeThisClassToo');
+  it('delete all data from the schema', () => {
+    const newClass: any = newClassObject('LetsDeleteThisClass');
+    const newClass2: any = newClassObject('LetsDeleteThisClassToo');
     const classNames = [newClass.class, newClass2.class];
     Promise.all([
       client.schema.classCreator().withClass(newClass).do(),
@@ -406,7 +406,7 @@ describe('schema', () => {
     ])
       .then(() => client.schema.getter().do())
       .then((schema) => classNames.forEach((cn) => expect(schema.classes?.map((c) => c.class)).toContain(cn)))
-      .then(() => client.schema.purge())
+      .then(() => client.schema.deleteAll())
       .then(() => client.schema.getter().do())
       .then((schema) =>
         classNames.forEach((cn) => expect(schema.classes?.map((c) => c.class)).not.toContain(cn))
